+++ sample/src/main/kotlin/ca/allanwang/kau/sample/AnimActivity.kt
@@ -49,14 +49,9 @@ class AnimActivity : KauBaseActivity() {
         }
     }
 
-    override fun onPostCreate(savedInstanceState: Bundle?) {
-        super.onPostCreate(savedInstanceState)
-        kauSwipeOnPostCreate()
-    }
-
     override fun onDestroy() {
-        super.onDestroy()
         kauSwipeOnDestroy()
+        super.onDestroy()
     }
 
     override fun onBackPressed() {