+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## v3.3.1
 * :core: Open up all logger functions
+* :core: Deprecate kauSwipeOnPostCreate and move functionality to onCreate
 * :searchview: Fix background tint
 
 ## v3.3.0